+++ all-is-cubes/src/block/block_def.rs
@@ -5,6 +5,7 @@
 )]
 
 use alloc::boxed::Box;
+use alloc::collections::BTreeMap;
 use alloc::sync::Arc;
 use core::{fmt, mem, ops};
 
@@ -31,21 +32,43 @@ use cache::{CacheUpdate, CachedBlock};
 
 // -------------------------------------------------------------------------------------------------
 
-/// Contains a [`Block`] and can be stored in a [`Universe`].
-/// Together with [`Primitive::Indirect`], this allows mutation of a block definition such
-/// that all its existing usages follow.
+/// A reusable, redefinable, animatable [`Block`] in a [`Universe`].
 ///
-/// To perform such a mutation, use [`BlockDefTransaction`].
+/// [`BlockDef`]s are used by being inserted into a [`Universe`], then creating blocks with
+/// [`Primitive::Indirect`] to refer to the definition. Reasons to use [`BlockDef`] include:
 ///
-/// Additionally, it caches the results of block evaluation to improve performance.
-/// Note that this cache only updates when the owning [`Universe`] is being stepped, or when
-/// a direct mutation to this [`BlockDef`] is performed, not when the contained [`Block`]
-/// sends a change notification.
+/// * Giving a [`Name`][universe::Name] to a block.
+///
+/// * Allowing cyclic relationships such as two or more blocks that turn into each other.
+///
+/// * Allowing redefinition of the block (using [`BlockDefTransaction`]),
+///   affecting all of its uses in the universe.
+///
+/// * [Animating][Self::new_animated] a block without the cost of each instance replacing itself
+///   separately.
+///
+/// * Caching the results of block evaluation to improve performance.
+///
+///   Note that this cache only updates when the owning [`Universe`] is being stepped, or when
+///   a direct mutation to this [`BlockDef`] is performed.
+///   Therefore, it is possible to receive stale evaluations (but change notifications will
+///   report when a fresh evaluation is available).
 pub struct BlockDef {
     state: BlockDefState,
 
     notifier: Arc<Notifier<BlockChange>>,
 }
+
+/// Looping animation of a [`Block`], defined as a set of [`Block`]s and the [phases][time::Phase]
+/// when they are shown.
+///
+/// Animations are used by putting them in [`BlockDef`]s;
+/// [`Animation`] itself is only needed when working with [`BlockDefTransaction`]s.
+/// It can be constructed by conversion `From<Block>`.
+///
+/// This type is not cheap to clone.
+#[derive(Clone, Eq, Hash, PartialEq)]
+pub struct Animation(BTreeMap<time::Phase, Block>);
 
 /// Component of [`BlockDef`] that is replaced when its definition is overwritten.
 ///
@@ -54,7 +77,17 @@ pub struct BlockDef {
 #[derive(ecs::Component)]
 #[require(CacheUpdate, NotifierComponent)]
 pub(crate) struct BlockDefState {
-    cached_block: CachedBlock,
+    /// Animation frames.
+    ///
+    /// At any particular time, this `BlockDef` exposes one frame as its current value.
+    /// Each frame is visible starting at the clock phase which is its key, up to the next
+    /// frame.
+    ///
+    /// Otherwise, there are two or more frames, none of which are necessarily at phase 0.
+    frames: BTreeMap<time::Phase, CachedBlock>,
+
+    /// The key of the frame in `frames` that is currently active.
+    current_frame: time::Phase,
 }
 
 /// Notifier of changes to this `BlockDef` entity's evaluation result, either via transaction or via
@@ -83,10 +116,42 @@ pub struct Read<'ticket> {
 impl BlockDef {
     /// Constructs a new [`BlockDef`] that stores the given block (which may be replaced
     /// in the future).
+    ///
+    /// This is equivalent to [`BlockDef::new_animated()`] with a single frame.
     pub fn new(read_ticket: ReadTicket<'_>, block: Block) -> Self {
         BlockDef {
+            state: BlockDefState::single(CachedBlock::new(block, read_ticket)),
+            notifier: Arc::new(Notifier::new()),
+        }
+    }
+
+    /// Constructs a new [`BlockDef`] that animates over time.
+    ///
+    /// Each element of `frames` is a pair of the block to display and the *first* clock phase at
+    /// which it is displayed.
+    ///
+    /// `frames` may be an [`Animation`] value, but this is neither required nor preferred.
+    #[track_caller]
+    pub fn new_animated(
+        read_ticket: ReadTicket<'_>,
+        frames: impl IntoIterator<Item = (time::Phase, Block)>,
+    ) -> Self {
+        let frames: BTreeMap<time::Phase, CachedBlock> = frames
+            .into_iter()
+            .map(|(phase, block)| (phase, CachedBlock::new(block, read_ticket)))
+            .collect();
+        assert!(
+            !frames.is_empty(),
+            "animated BlockDef must have at least one frame"
+        );
+        BlockDef {
             state: BlockDefState {
-                cached_block: CachedBlock::new(block, read_ticket),
+                current_frame: if frames.contains_key(&0) {
+                    0
+                } else {
+                    *frames.keys().last().unwrap()
+                },
+                frames,
             },
             notifier: Arc::new(Notifier::new()),
         }
@@ -94,14 +159,19 @@ impl BlockDef {
 
     /// Returns the current block value.
     ///
+    /// If the [`BlockDef`] is animated, returns the frame for phase 0 of the animation.
+    // NOTE: This is because a `BlockDef` not in a `Universe` can't be stepped.
+    ///
     /// Note that if you wish to get the [`EvaluatedBlock`] result, you should obtain the cached
     /// value by calling [`BlockDef::evaluate()`], or by using a [`Primitive::Indirect`],
     /// not by calling `.block().evaluate()`, which is not cached.
     pub fn block(&self) -> &Block {
-        self.state.cached_block.block()
+        self.state.current_frame().block()
     }
 
     /// Returns the current cached evaluation of the current block value.
+    ///
+    /// If the [`BlockDef`] is animated, returns the evaluation for phase 0 of the animation.
     ///
     /// This returns the same success or error as `Block::from(handle_to_self).evaluate()` would,
     /// not the same as `.block().evaluate()` would.
@@ -125,29 +195,32 @@ impl BlockDef {
 
 impl fmt::Debug for BlockDef {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        // TODO: Consider printing the cache, but only if it wouldn't be redundant?
-        let Self {
-            state: BlockDefState { cached_block },
-            notifier,
-        } = self;
-        f.debug_struct("BlockDef")
-            .field("cached_block", &cached_block)
-            .field("notifier", &notifier)
-            .finish()
+        universe::SealedMember::read_from_standalone(self).fmt(f)
     }
 }
 
 impl fmt::Debug for Read<'_> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        // TODO: Consider printing the cache, but only if it wouldn't be redundant?
         let Self {
-            state: BlockDefState { cached_block },
+            state:
+                BlockDefState {
+                    frames,
+                    current_frame,
+                },
             notifier,
         } = self;
-        f.debug_struct("BlockDef")
-            .field("cached_block", &cached_block)
-            .field("notifier", &notifier)
-            .finish()
+        let mut ds = f.debug_struct("BlockDef");
+        if frames.len() == 1
+            && *current_frame == 0
+            && let Some(frame) = frames.get(&0)
+        {
+            ds.field("cached_block", frame);
+        } else {
+            ds.field("frames", frames);
+            ds.field("current_frame", current_frame);
+        }
+        ds.field("notifier", notifier);
+        ds.finish()
     }
 }
 
@@ -173,23 +246,10 @@ impl listen::Listen for Read<'_> {
     }
 }
 
-impl AsRef<Block> for BlockDef {
-    fn as_ref(&self) -> &Block {
-        self.block()
-    }
-}
-
 impl VisitHandles for BlockDef {
     fn visit_handles(&self, visitor: &mut dyn HandleVisitor) {
         let Self { state, notifier: _ } = self;
         state.visit_handles(visitor);
-    }
-}
-
-impl VisitHandles for BlockDefState {
-    fn visit_handles(&self, visitor: &mut dyn HandleVisitor) {
-        let Self { cached_block } = self;
-        cached_block.visit_handles(visitor);
     }
 }
 
@@ -224,7 +284,11 @@ impl universe::SealedMember for BlockDef {
         })
     }
 
-    fn into_bundle(value: Box<Self>, _context: &universe::IntoBundleContext<'_>) -> Self::Bundle {
+    fn into_bundle(
+        mut value: Box<Self>,
+        context: &universe::IntoBundleContext<'_>,
+    ) -> Self::Bundle {
+        value.state.current_frame = find_frame(&value.state.frames, context.clock.phase()).0;
         (value.state, NotifierComponent(value.notifier))
     }
 }
@@ -251,17 +315,52 @@ impl<'a> arbitrary::Arbitrary<'a> for BlockDef {
 
 // -------------------------------------------------------------------------------------------------
 
+impl BlockDefState {
+    fn single(cached_block: CachedBlock) -> Self {
+        Self {
+            frames: BTreeMap::from([(0, cached_block)]),
+            current_frame: 0,
+        }
+    }
+
+    fn current_frame(&self) -> &CachedBlock {
+        &self.frames[&self.current_frame]
+    }
+
+    fn current_frame_mut(&mut self) -> &mut CachedBlock {
+        self.frames.get_mut(&self.current_frame).unwrap()
+    }
+}
+
+impl VisitHandles for BlockDefState {
+    fn visit_handles(&self, visitor: &mut dyn HandleVisitor) {
+        let Self {
+            frames,
+            current_frame: 0..,
+        } = self;
+        frames.visit_handles(visitor);
+    }
+}
+
+// -------------------------------------------------------------------------------------------------
+
 impl Read<'_> {
     /// Returns the current block value.
+    ///
+    /// If this is an animated [`BlockDef`], then the block returned is the current frame of that
+    /// animation.
     ///
     /// Note that if you wish to get the [`EvaluatedBlock`] result, you should obtain the cached
     /// value by calling [`BlockDef::evaluate()`], or by using a [`Primitive::Indirect`],
     /// not by calling `.block().evaluate()`, which is not cached.
     pub fn block(&self) -> &Block {
-        self.state.cached_block.block()
+        self.state.current_frame().block()
     }
 
     /// Returns the current cached evaluation of the current block value.
+    ///
+    /// If this is an animated [`BlockDef`], then the evaluation returned is the current frame of
+    /// that animation.
     ///
     /// This returns the same success or error as `Block::from(handle_to_self).evaluate()` would,
     /// not the same as `.block().evaluate()` would.
@@ -301,54 +400,148 @@ impl Read<'_> {
             self.notifier.listen_raw(listener.clone());
         }
 
+        // TODO: The returned evaluation should have an `AnimationHint` set, but that needs more
+        // cleverness to work well with the cache.
+
         if skip_eval {
             // In this case, don't use the cache, because it might contain an error, which
             // would imply the *listen* part also failed, which it did not.
             Ok(block::AIR_EVALUATED_MIN)
         } else {
-            self.state.cached_block.cached_evaluation()
+            self.state.current_frame().cached_evaluation()
         }
     }
 }
 
 // -------------------------------------------------------------------------------------------------
+// Impls for `Animation`
+
+impl Animation {
+    /// Creates an [`Animation`] value.
+    ///
+    /// Panics if `frames` does not contain at least one frame.
+    #[track_caller]
+    pub fn new<T: IntoIterator<Item = (time::Phase, Block)>>(frames: T) -> Self {
+        let map = BTreeMap::from_iter(frames);
+        assert!(
+            !map.is_empty(),
+            "animated BlockDef must have at least one frame"
+        );
+        Animation(map)
+    }
+
+    /// Returns an iterator over the frames of this animation, in ascending order.
+    ///
+    /// Note that the first element is not necessarily the frame which displays at phase 0.
+    pub fn iter(&self) -> impl Iterator<Item = (time::Phase, &Block)> {
+        self.0.iter().map(|(&k, v)| (k, v))
+    }
+
+    /// Construct the [`BlockDefState`] for this animation, evaluating the blocks.
+    fn into_state(self, current_phase: time::Phase, read_ticket: ReadTicket<'_>) -> BlockDefState {
+        // TODO: should we parallelize the evaluations?
+        let frames = self
+            .0
+            .into_iter()
+            .map(|(phase, block)| (phase, CachedBlock::new(block, read_ticket)))
+            .collect();
+        BlockDefState {
+            current_frame: find_frame(&frames, current_phase).0,
+            frames,
+        }
+    }
+}
+
+impl fmt::Debug for Animation {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        self.0.fmt(f)
+    }
+}
+
+impl From<Block> for Animation {
+    fn from(block: Block) -> Self {
+        Animation::new([(0, block)])
+    }
+}
+
+// Allows comparison inside transactions.
+// This impl is not public because `CachedBlock` isn't.
+impl PartialEq<BTreeMap<time::Phase, CachedBlock>> for Animation {
+    fn eq(&self, other: &BTreeMap<time::Phase, CachedBlock>) -> bool {
+        itertools::equal(
+            self.0.iter(),
+            other.iter().map(|(phase, cb)| (phase, cb.block())),
+        )
+    }
+}
+
+impl IntoIterator for Animation {
+    type Item = (time::Phase, Block);
+
+    type IntoIter = alloc::collections::btree_map::IntoIter<time::Phase, Block>;
+
+    /// Returns an iterator over the frames of this animation, in ascending order.
+    ///
+    /// Note that the first element is not necessarily the frame which displays at phase 0.
+    fn into_iter(self) -> Self::IntoIter {
+        self.0.into_iter()
+    }
+}
+
+/// Find which animation frame applies to the given clock phase.
+/// `frames` must be non-empty.
+fn find_frame<T>(frames: &BTreeMap<time::Phase, T>, phase: time::Phase) -> (time::Phase, &T) {
+    let (&key, value) = frames
+        .range(0..=phase)
+        .next_back()
+        .or_else(|| {
+            // If there is no frame with key <= phase then it must be the case that we're
+            // wrapping around and the active frame is the largest-numbered frame.
+            frames.last_key_value()
+        })
+        .expect("animation frame list should never be empty");
+    (key, value)
+}
+
+// -------------------------------------------------------------------------------------------------
 // Transaction for modifying `BlockDef`
 
-/// A [`Transaction`] which replaces (or checks) the [`Block`] stored in a [`BlockDef`].
+/// A [`Transaction`] which replaces (or checks) the [`Block`] or animation frames stored in a
+/// [`BlockDef`].
 #[derive(Clone, Debug, Default, Eq, Hash, PartialEq)]
 #[must_use]
 pub struct BlockDefTransaction {
-    // TODO: This struct is the second occurrence (the first is space::CubeTransaction) of a "assign to a mutable location" transaction. If we figure out how to have conveniently _composable_ transactions then we should have an `impl Transaction<Target = &mut T> for Assign<T>` transaction (targeting `&mut` to discourage use otherwise).
-    /// Block that must already be present.
-    old: Equal<Block>,
-    /// Block to be written in.
-    new: Equal<Block>,
+    /// Block or animation frames that must already be present.
+    old: Equal<Animation>,
+    /// Block or animation frames to replace the `BlockDefState` with when this transaction is
+    /// committed.
+    new: Equal<Animation>,
 }
 
 impl BlockDefTransaction {
     /// Returns a transaction which fails if the current value of the [`BlockDef`] is not
     /// equal to `old`.
-    pub fn expect(old: Block) -> Self {
+    pub fn expect(old: impl Into<Animation>) -> Self {
         Self {
-            old: Equal(Some(old)),
+            old: Equal(Some(old.into())),
             new: Equal(None),
         }
     }
 
     /// Returns a transaction which replaces the current value of the [`BlockDef`] with `new`.
-    pub fn overwrite(new: Block) -> Self {
+    pub fn overwrite(new: impl Into<Animation>) -> Self {
         Self {
             old: Equal(None),
-            new: Equal(Some(new)),
+            new: Equal(Some(new.into())),
         }
     }
 
     /// Returns a transaction which replaces the value of the [`BlockDef`] with `new`,
     /// if it is equal to `old`, and otherwise fails.
-    pub fn replace(old: Block, new: Block) -> Self {
+    pub fn replace(old: impl Into<Animation>, new: impl Into<Animation>) -> Self {
         Self {
-            old: Equal(Some(old)),
-            new: Equal(Some(new)),
+            old: Equal(Some(old.into())),
+            new: Equal(Some(new.into())),
         }
     }
 }
@@ -375,9 +568,7 @@ impl Transaction for BlockDefTransaction {
         read_ticket: Self::Context<'_>,
     ) -> Result<Self::CommitCheck, Self::Mismatch> {
         // Check the transaction precondition
-        self.old
-            .check(target.state.cached_block.block())
-            .map_err(|_| BlockDefMismatch::Unexpected)?;
+        self.old.check(&target.state.frames).map_err(|_| BlockDefMismatch::Unexpected)?;
 
         // Perform evaluation now, while we have the read ticket.
         //
@@ -386,8 +577,10 @@ impl Transaction for BlockDefTransaction {
         //
         // TODO: Maybe add an option to the transaction to be strict?
         Ok(Check {
-            new_state: self.new.0.clone().map(|block| BlockDefState {
-                cached_block: CachedBlock::new(block, read_ticket),
+            new_state: self.new.0.clone().map(|animation| {
+                // Note that using current_frame is lossy, if the old animation had fewer
+                // frames. Ideally we would also have the universe clock phase unaltered.
+                animation.into_state(target.state.current_frame, read_ticket)
             }),
         })
     }
@@ -419,9 +612,7 @@ impl universe::TransactionOnEcs for BlockDefTransaction {
         read_ticket: ReadTicket<'_>,
     ) -> Result<Self::CommitCheck, Self::Mismatch> {
         // Check the transaction precondition
-        self.old
-            .check(target.state.cached_block.block())
-            .map_err(|_| BlockDefMismatch::Unexpected)?;
+        self.old.check(&target.state.frames).map_err(|_| BlockDefMismatch::Unexpected)?;
 
         // Perform evaluation now, while we have the read ticket.
         //
@@ -430,8 +621,10 @@ impl universe::TransactionOnEcs for BlockDefTransaction {
         //
         // TODO: Maybe add an option to the transaction to be strict?
         Ok(Check {
-            new_state: self.new.0.clone().map(|block| BlockDefState {
-                cached_block: CachedBlock::new(block, read_ticket),
+            new_state: self.new.0.clone().map(|animation| {
+                // Note that using current_frame is lossy, if the old animation had fewer
+                // frames. Ideally we would also have the universe clock phase unaltered.
+                animation.into_state(target.state.current_frame, read_ticket)
             }),
         })
     }
@@ -528,6 +721,8 @@ impl fmt::Display for BlockDefConflict {
     }
 }
 
+// -------------------------------------------------------------------------------------------------
+
 #[derive(Clone, Copy, Debug, Default, PartialEq)]
 pub(crate) struct BlockDefStepInfo {
     /// A cache update was attempted.
@@ -582,8 +777,25 @@ pub(crate) fn add_block_def_systems(world: &mut ecs::World) {
     let mut schedules = world.resource_mut::<ecs::Schedules>();
     schedules.add_systems(
         time::schedule::Synchronize,
-        (update_phase_1, update_phase_2).chain().in_set(BlockDefUpdateSet),
+        (animate, update_phase_1, update_phase_2).chain().in_set(BlockDefUpdateSet),
     );
+}
+
+/// ECS system function that advances [`BlockDef`] animations.
+fn animate(
+    current_step: ecs::Res<universe::CurrentStep>,
+    mut defs: ecs::Query<(&mut BlockDefState, &NotifierComponent)>,
+) -> ecs::Result {
+    let current_phase = current_step.get()?.tick.next_phase();
+    for (mut def_state, NotifierComponent(notifier)) in defs.iter_mut() {
+        let (new_frame, _) = find_frame(&def_state.frames, current_phase);
+
+        if def_state.current_frame != new_frame {
+            def_state.current_frame = new_frame;
+            notifier.notify(&BlockChange::new());
+        }
+    }
+    Ok(())
 }
 
 /// ECS system function that looks for `BlockDef`s needing reevaluation, then writes the new
@@ -604,7 +816,9 @@ fn update_phase_1(
             "CacheUpdate should have been cleared",
         );
 
-        if let Some(update) = def_state.cached_block.update_phase_1(&mut info, read_ticket) {
+        // TODO: Instead of checking only the current frame, check other frames
+        // (maybe only if free time permits).
+        if let Some(update) = def_state.current_frame().update_phase_1(&mut info, read_ticket) {
             *next = update;
         }
     }
@@ -628,8 +842,9 @@ fn update_phase_2(
             // By bypassing change detection, we avoid detecting this consumption of the change.
             // (This means that change detection no longer strictly functions as change detection,
             // but that is okay because `CacheUpdate` is *only* for this.)
-            let changed =
-                def_state.cached_block.update_phase_2(mem::take(next.bypass_change_detection()));
+            let changed = def_state
+                .current_frame_mut()
+                .update_phase_2(mem::take(next.bypass_change_detection()));
             if changed {
                 notifier.notify(&BlockChange::new());
             }
@@ -641,14 +856,15 @@ fn update_phase_2(
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::content::make_some_blocks;
     use crate::math::Rgba;
     use crate::universe::Universe;
     use pretty_assertions::assert_eq;
 
-    /// Quick more-than-nothing test for [`BlockDef::evaluate()`] being the same as more usual
-    /// options.
+    /// Test [`BlockDef::evaluate()`] and [`Read::evaluate()`] being the same as
+    /// [`Primitive::Indirect`].
     ///
-    /// TODO: Test its behavior on failure.
+    /// TODO: Test its behavior on evaluation failures.
     #[test]
     fn evaluate_equivalence() {
         let mut universe = Universe::new();
@@ -685,4 +901,148 @@ mod tests {
             "BlockDef::evaluate() same except for block and cost as the def block"
         );
     }
+
+    fn animation_test_universe() -> Box<Universe> {
+        let mut universe = Universe::new();
+        universe.set_clock(time::Clock::new(time::TickSchedule::per_second(4), 0));
+        universe
+    }
+
+    /// Steps the universe for `expected_frames.len() - 1` steps.
+    fn run_and_assert_frames(
+        universe: &mut Universe,
+        block_def_handle: &universe::Handle<BlockDef>,
+        expected_frames: &[&Block],
+    ) {
+        let change_flag = listen::Flag::listening(
+            false,
+            &block_def_handle.read(universe.read_ticket()).unwrap(),
+        );
+
+        let mut previous_block: Option<Block> = None;
+        for (i, expected_block) in expected_frames.iter().copied().enumerate() {
+            // Step, but only between iterations, not before the first.
+            if i != 0 {
+                println!("step #{i}/{}", expected_frames.len() - 1);
+                universe.step(false, time::Deadline::Whenever);
+            }
+
+            // Check current block value
+            {
+                let read_block_def = block_def_handle.read(universe.read_ticket()).unwrap();
+                println!(
+                    "at phase {} got block {:?}",
+                    universe.clock().phase(),
+                    read_block_def.evaluate().unwrap().attributes().display_name
+                );
+                assert_eq!(
+                    read_block_def.block(),
+                    expected_block,
+                    "read().block() should be the expected frame"
+                );
+            }
+
+            // Check for a change notification
+            let got_notification = change_flag.get_and_clear();
+            assert_eq!(
+                got_notification,
+                previous_block.is_some_and(|pb| pb != *expected_block),
+                "notification should be present iff the block is different"
+            );
+            previous_block = Some(expected_block.clone());
+        }
+    }
+
+    #[test]
+    #[should_panic = "animated BlockDef must have at least one frame"]
+    fn animation_type_not_empty() {
+        Animation::new([]);
+    }
+
+    #[test]
+    #[should_panic = "animated BlockDef must have at least one frame"]
+    fn new_animated_not_empty() {
+        BlockDef::new_animated(ReadTicket::stub(), []);
+    }
+
+    #[test]
+    fn animation_from_new_animated() {
+        let mut universe = animation_test_universe();
+        let [frame_a, frame_b] = make_some_blocks();
+        // TODO: test insertion into a universe with a nonzero phase, too
+        let block_def_handle = universe
+            .insert(
+                "animated".into(),
+                BlockDef::new_animated(
+                    universe.read_ticket(),
+                    // first frame not being at phase 0 exercises wrapping
+                    Animation::new([(1, frame_a.clone()), (3, frame_b.clone())]),
+                ),
+            )
+            .unwrap();
+
+        run_and_assert_frames(
+            &mut universe,
+            &block_def_handle,
+            &[&frame_b, &frame_a, &frame_a, &frame_b, &frame_b],
+        );
+    }
+
+    #[test]
+    fn animation_initial_phase_not_zero() {
+        let mut universe = animation_test_universe();
+        universe.step(false, time::Deadline::Whenever); // phase now 1
+
+        let [frame_a, frame_b] = make_some_blocks();
+        let block_def_handle = universe
+            .insert(
+                "animated".into(),
+                BlockDef::new_animated(
+                    universe.read_ticket(),
+                    Animation::new([(0, frame_a.clone()), (1, frame_b.clone())]),
+                ),
+            )
+            .unwrap();
+
+        run_and_assert_frames(
+            &mut universe,
+            &block_def_handle,
+            &[&frame_b, &frame_b, &frame_b, &frame_a, &frame_b],
+        );
+    }
+
+    #[test]
+    fn replace_animation_with_animation() {
+        let mut universe = animation_test_universe();
+        let [frame_a, frame_b, frame_c, frame_d] = make_some_blocks();
+        let block_def_handle = universe
+            .insert(
+                "animated".into(),
+                BlockDef::new_animated(
+                    universe.read_ticket(),
+                    Animation::new([(1, frame_a.clone()), (3, frame_b.clone())]),
+                ),
+            )
+            .unwrap();
+
+        run_and_assert_frames(&mut universe, &block_def_handle, &[&frame_b, &frame_a]);
+
+        // Replace the animation with one with different frames and phases.
+        universe
+            .execute_1(
+                &block_def_handle,
+                BlockDefTransaction::overwrite(Animation::new([
+                    (0, frame_c.clone()),
+                    (2, frame_d.clone()),
+                ])),
+            )
+            .unwrap();
+        run_and_assert_frames(
+            &mut universe,
+            &block_def_handle,
+            &[&frame_c, &frame_d, &frame_d, &frame_c, &frame_c],
+        );
+    }
+
+    // TODO: Test animation with a frame number greater than clock divisor
 }